+++ src/Clkd.cpp
@@ -900,6 +900,7 @@ struct ClkdWidget : ModuleWidget {
 		apItem->idPtr = &module->id;
 		apItem->resetClockOutputsHighPtr = &module->resetClockOutputsHigh;
 		apItem->slaveResetRunBpmInputs = slaveResetRunBpmInputs;
+		apItem->slaveBpmDetectionMode = &module->bpmDetectionMode;
 		menu->addChild(apItem);
 	}
 	
@@ -1051,7 +1052,7 @@ struct ClkdWidget : ModuleWidget {
 			else if ( e.key == GLFW_KEY_M && ((e.mods & RACK_MOD_MASK) == RACK_MOD_CTRL) ) {
 				Clkd *module = dynamic_cast<Clkd*>(this->module);
 				if (clockMaster.id != module->id && clockMaster.validateClockModule()) {
-					autopatch(slaveResetRunBpmInputs, &module->resetClockOutputsHigh);
+					autopatch(slaveResetRunBpmInputs, &module->bpmDetectionMode, &module->resetClockOutputsHigh);
 				}
 				e.consume(this);
 				return;